+++ packages/apps/project/src/api/project-task-api.js
@@ -40,7 +40,7 @@ const createProjectTaskCriteria = ({
 
   if (userId != null) {
     criteria.push({
-      fieldName: 'assignedTo.employee.user.id',
+      fieldName: 'assignedTo.id',
       operator: '=',
       value: userId,
     });
