+++ packages/apps/project/src/components/molecules/index.ts
@@ -22,3 +22,4 @@ export {default as PartnerActionCard} from './PartnerActionCard/PartnerActionCar
 export {default as ProjectHeader} from './ProjectHeader/ProjectHeader';
 export {default as ProjectSiteTags} from './ProjectSiteTags/ProjectSiteTags';
 export {default as TaskActionCard} from './TaskActionCard/TaskActionCard';
+export {default as TaskHeader} from './TaskHeader/TaskHeader';
